+++ toolbox/mdcode/tests/libts/semantic/fixtures/sales_bq_graph_target.osi.golden.yaml
@@ -1,0 +1,29 @@
+version: 0.2.0.dev0
+semantic_model:
+  - name: sales
+    custom_extensions:
+      - vendor_name: GOOGLE
+        data: '{"deploymentTargets":
+          ["//bigquery.googleapis.com/projects/demo/datasets/sales/propertyGraphs/sales_graph"]}'
+    datasets:
+      - name: orders
+        source: demo.sales.orders
+        primary_key:
+          - o_orderkey
+        fields:
+          - name: o_orderkey
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_orderkey
+          - name: o_totalprice
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_totalprice
+    metrics:
+      - name: total_revenue
+        expression:
+          dialects:
+            - dialect: BIGQUERY
+              expression: SUM(orders.o_totalprice)
