+++ toolbox/mdcode/tests/libts/semantic/fixtures/star_orders_customer.pull.golden.yaml
@@ -1,0 +1,61 @@
+# (no warnings)
+version: 0.2.0.dev0
+semantic_model:
+  - name: sales
+    description: Sales orders with customer attributes
+    datasets:
+      - name: orders
+        source: samples.tpch.orders
+        description: One row per order
+        fields:
+          - name: o_orderkey
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_orderkey
+            description: Order identifier
+          - name: o_custkey
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_custkey
+          - name: o_orderdate
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_orderdate
+            dimension: {}
+          - name: o_totalprice
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: o_totalprice
+      - name: customer
+        source: samples.tpch.customer
+        fields:
+          - name: c_custkey
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: c_custkey
+          - name: c_name
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: c_name
+            description: Customer name
+    metrics:
+      - name: total_revenue
+        expression:
+          dialects:
+            - dialect: BIGQUERY
+              expression: SUM(orders.o_totalprice)
+        datatype: Decimal
+        description: Total order revenue
+      - name: order_count
+        expression:
+          dialects:
+            - dialect: BIGQUERY
+              expression: COUNT(orders.o_orderkey)
+        datatype: Decimal
+        description: Number of orders
