+++ toolbox/mdcode/tests/libts/semantic/fixtures/tpcds_date_edge.pull.golden.yaml
@@ -1,0 +1,147 @@
+# (no warnings)
+version: 0.2.0.dev0
+semantic_model:
+  - name: tpcds_model
+    description: TPC-DS retail model
+    datasets:
+      - name: store_sales
+        source: tpcds.public.store_sales
+        description: Fact table containing all store sales transactions
+        fields:
+          - name: ss_item_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_item_sk
+            dimension: {}
+          - name: ss_ticket_number
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_ticket_number
+            dimension: {}
+          - name: ss_customer_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_customer_sk
+            dimension: {}
+          - name: ss_store_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_store_sk
+            dimension: {}
+          - name: ss_quantity
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_quantity
+            description: Quantity of items sold
+          - name: ss_sales_price
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_sales_price
+            description: Sales price per unit
+          - name: ss_ext_sales_price
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_ext_sales_price
+            description: Extended sales price (quantity * price)
+          - name: ss_net_profit
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: ss_net_profit
+            description: Net profit from the sale
+      - name: customer
+        source: tpcds.public.customer
+        description: Customer dimension with demographic information
+        fields:
+          - name: c_customer_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: c_customer_sk
+            dimension: {}
+          - name: c_first_name
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: c_first_name
+            dimension: {}
+            description: Customer first name
+          - name: c_last_name
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: c_last_name
+            dimension: {}
+            description: Customer last name
+      - name: item
+        source: tpcds.public.item
+        description: Item/Product dimension
+        fields:
+          - name: i_item_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: i_item_sk
+            dimension: {}
+          - name: i_brand
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: i_brand
+            dimension: {}
+          - name: i_category
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: i_category
+            dimension: {}
+          - name: i_current_price
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: i_current_price
+            description: Current price of the item
+      - name: store
+        source: tpcds.public.store
+        description: Store dimension with location attributes
+        fields:
+          - name: s_store_sk
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: s_store_sk
+            dimension: {}
+          - name: s_store_name
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: s_store_name
+            dimension: {}
+          - name: s_city
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: s_city
+            dimension: {}
+          - name: s_state
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: s_state
+            dimension: {}
+          - name: s_number_employees
+            expression:
+              dialects:
+                - dialect: BIGQUERY
+                  expression: s_number_employees
+            description: Number of employees at the store
+      - name: date_dim
+        source: sqlgen-testing.demo.date_dim
+        description: Date dimension with calendar attributes
